productId Integer로 변경

+++ src/main/java/com/fastquick/data/dto/response/ProductListResponseDTO.java
@@ -4,12 +4,12 @@ import lombok.*;
 
 @Getter
 public class ProductListResponseDTO {
-	private Long productId;
+	private Integer productId;
     private String productName;
     private Long barcode;
     private Integer quantity;
     
-    public ProductListResponseDTO(Long productId, String productName, Long barcode, Integer quantity) {
+    public ProductListResponseDTO(Integer productId, String productName, Long barcode, Integer quantity) {
     	this.productId = productId;
     	this.productName = productName;
     	this.barcode = barcode;
